Rounded Rectangle Background Support - Rafactor
+Added function de render uielement background and support the rounded rectangle

+++ SFML-UIAddon/SFML-UIAddon-Dll/Code/Headers/SFUIL/Graphics/RoundedRectangle.hpp
@@ -20,9 +20,9 @@ namespace sfui
 
         float getRadius();
 
-        void setPointsPerCorner(float _points);
+        void setPointsPerCorner(size_t _points);
 
-        int getPointsPerCorner();
+        size_t getPointsPerCorner();
 
     private:
         size_t m_pointsPerCorner = 2;
